- Added install commands to CMake files for MSVC_IDE. This almost completely fixes the problem where plugins' DLL's would be built into wrong directory and left there. The remaining problem is that INSTALL project in the MSVC solution is disabled by default... - Added header files to CMake file lists so proper projects can now be generated for IDEs like Code::Blocks or Microsoft Visual Studio. (NOTE: Some files may still be missing, the project builds fine in MSVC and MinGW anyway) - Fixed some particularly annoying warnings generated by MSVC compiler.

+++ src/sdeapi/config.cpp
@@ -187,7 +187,7 @@ void Config::readConfig()
 				if(!sc.checkToken(TK_IntConst))
 					break;
 				createSetting(index, 0);
-				setting(index)->setValue(negative ? -sc.number : sc.number);
+				setting(index)->setValue(negative ? -(int)(sc.number) : sc.number);
 			}
 			if(!sc.checkToken(';'))
 				break;
